Consider minimized window as visible ones.
In real minimized windows are hidden but with additional flag "minimized". To distinguish between completely hidden window and just minimized window we check flag "minimized" additionally. Only if flag "hidden" is set and "minimized" flag is not set the window is recognized as hidden by application. Without this differentiation also minimized window will not show up in windows view and will not be selectable or restorable to previous state (a visible state) anymore. At least not via xfdashboard but via task switcher (ALT+Tab).