Fix bug that XfdashboardCollapseBox collapse even if pointer device is still inside actor
There are a few situations where only a "leave-event" signal without a following "enter-event" signal is emitted although the pointer is still inside the XfdashboardCollapseBox actor. So the actor is shrinked to minimum size accidently. This happens most when switching through workspaces. So we first check if the new target actor is a direct or deeper child of the collapse box and if it is not then we check if the actor under pointer device is a direct or deeper child. If both checks fails then the pointer device should be really outside the collapse box and it is allowed to shrink to minimum size.