Add theming support
Added support for theming xfdashboard. In addition it is now a requirement that least one theme is installed and referenced in xfconf at property /theme. A default theme named "xfdashboard" is included and used if xfconf property is not set. This theme is a copy of the previously hard coded style.