Fix visual bug when drawing outlines of XfdashboardBackground derived actors
The outline was drawn when virtual paint_node() function of XfdashboardBackground was drawn. Any XfdashboardBackground derived object could easily overdraw outline. Outline is now a effect which takes care that outline is drawn last and is always visible. Also a new property "outline-borders" (XfdashboardBackground) which "maps" to "borders" in new XfdashboardOutlineEffect class and manages which sides (and also which rounded corner) should be drawn.