Register our GValue transformation function on application initialization
Move GValue transformation function to utils.c and created a public register function which registers all our transformation function. This register function is called when the first application instance is created to get these transformation function available at the earliest time because they might be needed not only in theming (in XfdashboardActor).