Rename key in theme file from "LayoutPrimary" to "Layout"
When writing documentation about how to layout xfdashboard I saw that is was not wise to name the key "LayoutPrimary". Because the key takes a list of file, separated by semicolon, there is no need to differ between layout files and their purposes. Only the attribute "id" of the object element which is the child of interface needs to set correctly.