Move stylable functions into an interface
XfdashboardActor will not implement and provide styling possibilities but implements the new stylable interface (XfdashboardStylable). Moving the functionality into an interface gives us the opportunity to style non-actors also (anytime later).

Why should non-actors be stylable by CSS? That's because non-actors like XfdashboardWindowContents or XfdashboardImage will be visible because they are used by (visible and drawn) actors. This is not implemented yet but prepared.