Rename class XfdashboardImage to XfdashboardImageContent and refactor code
Warning: Massive API break because of renamed class and functions. Renamed XfdashboardImage to XfdashboardImageContent to keep class name in sync with naming convention I used so far (see XfdashboardWindowContent for a ClutterContent derived class and ClutterImage, which XfdashboardImage is derived from, is also a ClutterContent derived class). Also XfdashboardImage is now free e.g. for an actor utilizing XfdashboardImageContent.