Add support to prevent upscaling beyond real size in XfdashboardScaledTableLayout
The layout manager XfdashboardScaledTableLayout got a new property called "prevent-upscaling" which is set to FALSE by default. This property manages if layout manager is allowed to scale a child larger than its real natural size or not. The windows view uses this new property and exports it by itself to allow windows to be scaled beyond their real size or not. As it is a stylable property it can be control by a theme.