Simplify checking visibility of a garcon menu element
Just call get_visible() on a GarconMenuElement as it will do already all extensive checks if this element should be visible or not. There is no need to do our additional checks because they are already done. But we still have to count the number of visible sub-menus and menu items to determine if we should show or hide a menu.