Try to create GDesktopAppInfo for GarconMenuItem first then fallback to previous method
We try first to create a GDesktopAppInfo for the requested GarconMenuItem. If it succeeds launching the application via this GDesktopAppInfo will behave as expected, e.g. setting working directory from "Path=..." of desktop file etc. If this creation fails fallback to old method with its restriction, e.g. not setting the working directory etc.