Add setting switch-to-view-on-resume to let user decide to which view to switch when resuming
If setting switch-to-view-on-resume is set to a known internal name of a view and "reset-search-on-resume" is set to TRUE then switch to this view when resuming. If this setting is not set at all or empty or an unknown internal name then do the default behaviour to switch to previous selected view. This addresses issue #20