Set default hint label in search-box if none was set
Check if a hint text was set and set default hint label if not. XfdashboardTextBox exports a read-only property "hint-text-set" for this purpose. If either an empty or non-empty (but not NULL) string is set as hint label this flag will be set TRUE. A NULL string will unset "hint-text-set" property, means set to FALSE.

To get the default hint label set in search-box just do _not_ set property "hint-text" in theme's layout file. To get an empty hint label specify an empty "hint-text" in theme's layout file. Otherwise set hint label to the text wanted.

This commit addresses issue #26