Allow some ClutterActor properties to be stylable
Allow the properties '{x,y}-expand' and '{x,y}-align' to be themable. This addresses issue #28 to avoid that a cell in XfdashboardDynamicTableLayout is always filled completely by child (and its background). To avoid this set 'x-expand' and/or 'y-expand' to 'false' - maybe also set '{x,y}-align' accordingly to position child in available allocation space (e.g. set to 'center').