Alpha color component in @rgba() is a factor between 0.0 and 1.0 and not a common color component
Fix alpha function parameter retrieval and calculation. It was a mistake to handle the alpha color component as a common color component which can be defined either an integer between 0 and 255 or as a percentage between 0% and 100%. The alpha color component is a factor and therefore must be a floating point number between 0.0 and 1.0.