Implement file @-functions in CSS
The CSS parser was extended with a file function called @try_icons. This functions take a variable number of strings as functions arguments and lookups each one in the order they are given. The first argument which can be lookuped successfully will be returned. If no none can be lookuped an empty string will be returned which causes xfdashboard to load the fallback icon (aka missing icon). Each argument can either be an absolute path, a relative path to theme's path or a (stock) icon name. This commit closes issue #9