Revert changes to _xfdashboard_image_content_setup_for_icon to allow empty icon names
Revert changes to _xfdashboard_image_content_setup_for_icon in XfdashboardImageContent which allows to fallback to fallback icon (aka missing image) if an empty string was set, e.g. by @try_icons in CSS. Empty string will now prevent loading any image as intended. To fallback to fallback icon in @try_icon you should declare a constant like @missing-icon with either the icon name to fallback to or with @try_icons(...) function with a list of fallback icons to try. This @missing-icon can be used in CSS with "@try_icons(..., ..., ..., @missing-icon);". This variant is used in xfdashboard's default theme.