Fix loading scalable images by icon name
Load images by icon name with scaling at requested size. If file to load is a bitmap it should already be at right size but if is scalable image (svg) is was loaded as a too small bitmap and then scaled resulting in an ugly image but now should look good and "right-scaled".