Avoid closing windows accidently with key <Backspace>
A window can accidently be closed when starting a search by typing and correcting more letters than entered with key <Backspace>. Because then the view switchs from search view back to windows view, sets the focus to a window in the windows and the additional <Backspace> key press will close it. Changed behaviour that the shift key has to be pressed also when using <Backspace>. <Delete> will still work as before.

This commit addresses issue #32.