Simplify and generalize selection handling
Implemented a generic and common selection handling framework in XfdashboardFocusable. This has the advantage that we do not need to shift the key focus around between nothing and search box's text-box. We can also implement more common key press/release event like HOME and END keys to select first or last selectable item at the current focussed actor. Currently the keys UP, DOWN, LEFT, RIGHT, HOME, END, ENTER are recognized by this framework. Still missing are PAGE_UP and PAGE_DOWN. The disadvantage is that "uncommon" key events like CTRL+ENTER have to be catched by signal handler connected to 'key-press-event" and 'key-release-event' at the actor in question like XfdashboardWindowsView.

The search view is the only actor/view not implementing this framework as the decision if search view should be completely rewritten was not yet made.

This is also preparation work for (maybe coming) plugins.