Move selection if current one gets unavailable
The new generic selection frameworks now recognizes if the current selection is getting unavailable because the current selected actor is going to be hidden or destroyed. Then it moves the selection to next available and selectable actor. Focusable actors do not need to handle destroyed actor which are selected by themselve anymore so remove this code logic (XfdashboardQuicklaunch and XfdashboardWindowsView).