Adding XML parser for effects to theme
First step in closing issue #34 by implementing a XML parser to theme which can read in XML data for effects. To get an effect XML file loaded the descriptive xfdashboard.theme file needs a new entry "Effects=" with a semi-colon separated list of files to load. The new XfdashboardThemeEffects object still misses the "real" creation function and also the basic XfdashboardActor does not make any use of it yet. But this is also only step one ;)