Implement effects into XfdashboardActor
XfdashboardActor has a new property 'effects' which takes a space-separated list of IDs of effects to create. The IDs must be found in effects XML file specified in "Effects=..." in theme description file xfdashboard.theme. The default theme defines a 'black-white' effect which decolorizes the actor and all quicklaunch buttons will get this effect added but the list is emptied at hovered button in quicklaunch. Effects can be expensive (seen from performance point of view) so changing effects on hover is a bad idea. So the default theme is currently a bad theme but a visible example was needed :(

This closes issue #34