Try not to eat up CPU
If a new window is created an instance of XfdashboardWindowContent is created and xfdashboard_window_content_set_window is called once for this new window. This function sets up the window content object to retrieve live updates. But it also sets up for live updates if xfdashboard is running in daemon mode and it is not active at this time. The result is that the window content texture is updated in background all the time until xfdashboard gets active and is suspended again then the window content's suspend function is called which stops listening for live update until xfdashboard is resumed.

This CPU usage is unneccessary so call window content's suspend function immediately in xfdashboard_window_content_set_window after it has been set up for live update to reduce CPU usage again.

I hope this fixes issue #44