Implement '@import' function in CSS
The @import function allows to import other CSS styles from other files. The syntax is: @import(filename). The filename can be an absolute or relative path. The filename should be quoted with either by single or double quotation marks but it must not. Relative paths will be looked up first relative to the path of current CSS file being parsed. If no such file exists the relative path will be used relative the theme's path.

This commit addresses issue #52