Adjust line of selector's source to a global line offset
This is needed because an imported CSS file via @import(...) will start at line 1 although it is included virtually into the current CSS file at the line where @import is issued. So the imported CSS file and all following lines in the current CSS file need to be adjusted. That is what the global line offset is for. This corrects sorting matching selectors when priority is equal and the line in file where the matching selector is checked (see _xfdashboard_theme_css_sort_by_score).