Ensure right size of window actor and window content
XfdashboardLiveWindow determines its size from XfdashboardWindowContent which either returns the size of the corresponding window if fallback is used or size of window content texture if fallback is not used. If application is suspended (and also all XfdashboardWindowContent objects) and it resumes the texture still returns the old size even if the window was resized in the meantime. This fix ensures that XfdashboardWindowContent returns window's real size if fallback is used (like before) or if it is suspended (that's new) otherwise it returns the size of window content texture (like before).

This fix addresses issue #54