Apply new theme while running application
This commit addresses and should fix issue #59. Rewrote XfdashboardApplication and XfdashboardStage to reload theme when xfconf property 'theme-name' changes. This introduces XfdashboardStageInterface as top-level actor for a stage (screen).

Sadly this also breaks current theme API. All references in CSS and XML files to class XfdashboardStage have to be replaced to XfdashboardStageInterface. The new actor XfdashboardStageInterface shares the same stylable properties as XfdashboardStage and all changes to XfdashboardStageInterface will be applied to XfdashboardStage on-the-fly. So XfdashboardStageInterface is a one-to-one in-place replacement for XfdashboardStage.

It is not the nicest solution as I wanted it to be but it could also be a preparation work for multi-monitor support (issue #14).