Implement key-bindings configuration
Implemented a new configuration option. The application uses now a dynamic keyboard binding configuration instead of hard-coded key bindings like key 'Left' for select the item left of the currently selected item in a focusable actor. The configuration file is a XML file describing type of binding (although only key bindings are supported yet), the keycode and modifiers for the binding and the source and the action signal to perform when this key combination (of keycode and modifiers) are pressed or released at the actor specified.

The application lookups up first a user-defined keyboard binding configuration file at 'xfdashboard/bindings.xml' the user's configuration folder which is ${HOME}/.config/xfdashboard/bindings.xml usually. If the file is not found at this path it looks up a system-wide keyboard configuration file which is a default one installed by xfdashboard, usually at /usr/share/xfdashboard/bindings.xml. The application will quit with error if none was found.

This commit is the first part for the issues #35 and #41