Add more bindable actions
Views can now be activated and focused (at once - simultaneously) but only if attribute "allow-unfocusable-targets" is set to true or 1 in bindings configuration file.

If attribute "allow-unfocusable-targets" is set to true or 1 it indicates that also unfocusable actor should be included when resolving targets. It should be used with caution because it needs the action function to set focusable state of actor to be true.