Allow scroll events in windows view to change active workspace
Like XfdashboardWorkspaceSelector the windows view should change active workspace on scroll event, e.g. mouse-wheel up or down. But it should be configurable so a new xfconf property "/components/windows-view/scroll-event-changes-workspace" was introduced which is a boolean value to indicate if it should change active workspace on scroll events (value is TRUE) or not (value is FALSE)..

This fixes issue #41