Allow icon of search provider in search view result container to be themable
This is achieved by adding two items.

The first one is a themable property named "icon" in XfdashboardSearchResultContainer which expects a string with a themed icon name or a filename of the icon to display.

The second one is a CSS class which will be added to XfdashboardSearchResultContainer. The CSS class name is build with a prefix "search-provider-" followed by the search provider's object name (like XfdashboardApplicationsSearchProvider for the applications search provider) - so the complete CSS class name would be ".search-provider-XfdashboardApplicationsSearchProvider" (for use in CSS) in this case.

This commit addresses issue #46