Add hotkey tracker
The hotkey tracker will watch all key events globally and toggle application state when hotkey was recognized. Toggling means suspending application if is currently active or resuming application if not. Currently the hotkey is haard-coded to left Super key and right Super key.

Reacting to hotkey presses is disabled by default. It has to be enabled by setting Xfconf's property "/hotkey-enabled" to boolean TRUE.

The hotkey tracker is only usable in daemon mode but works also when running standalone. In the last case it will quit xfdashboard if hotkey is pressed but it cannot activate xfdashboard (because therefore it needs to be running in daemon mode otherwise the needed code is not running ;) )