Fix build system
Respect configure options --[enable/disable]-xcomposite and --[enable/disable]-xdamage by checking if X11 extensions exist _AND_ set variables for Makefile to include necessary include paths and needed libs depending on the configure option.

Before I just checked the existence of required packages but nothing more. This means that enabling any of both extension was requested at 'configure' but clutter was not compiled with this extension the build must fail because the required header files and libraries were not be include because I did not add them in Makefile explicitly. I was relying on Clutter to be built with this extension and pulling them in.