Impleting XfdashboardDesktopAppInfo
XfdashboardDesktopAppInfo is a new class wrapping a GarconMenuItem and a string for the desktop ID and it also implements the GAppInfo interface. So it can be used - more or less - as a drop-in replacement for GDesktopAppInfo. Each function of GAppInfo interface is implemented except for function regarding to mime type associations. This class wraps GarconMenuItem to gain all it's advantages over GDesktopAppInfo, e.g. supporting desktop file of type "Link" to start a web-browser with the url specified in desktop file. It also forwards the "changed" signal of GarconMenuItem (if it did set up a file monitor for it) to notify about changes to the menu item and the desktop file. So an actor could adjust its appearing to the changes made.

This class also has an advance lookup function for desktop IDs containing dashes. It should find a desktop file for nearly any possible combination of a dash being a directory separator or being part of the file/directory name.

This commit addresses isse #67