Replace dependency on libwnck-1.0 (minimum version 2.30) with libwnck-3.0 (minimum version 3.0)
We depend on GTK+3 (version 3.2 or higher) in settings application and for some GTK+ functions in xfdashboard but we also depend on libwnck for window handling. Using libwnck-1.0 triggers to include GDK2 from GTK+2 resulting in a mix of GTK+ 2 and 3 what is really bad although it seems to work pretty well. In order to not mix both major GTK+ and to depend only on one GTK+ version the minimum version of libwnck had to raised up to 3.0 and since then we had to depend on libwnck-3.0 instead of libwnck-1.0.