Always check if window is on workspace in windows view
Previously it was check if a workspace is set at windows view and if it was then it was check if windows is on workspace. If a workspace was not set it was assumed that window is visible. This resulted in creating a window actor for each window at start-up which. After start-up a workspace was set which caused all window actor to be destroyed and for all windows which are now on the workspace set were recreated. This delayed start-up because lots of window actor had to be created twice.

Now we assume a window to be hidden if a workspace is not set because if windows view don't know the workspace no window could match it. This reduces start-up time a lot because no window actor is created twice but only once and creating window textures is a time consuming task.