Implement showing desktop also in XfdashboardLiveWorkspace
Added a themable property 'background-image-type' which behave like the property 'background-image-type' at XfdashboardStage(Interface). If set to 'desktop' it will show the desktop's background at the workspace (e.g. as used in workspace selector).

It is enabled by default as it is also enabled by default at XfdashboardStageInterface.

This commit addresses issue GH #81