Implement signal to notify if a window has moved to another monitor
Implement a new signal 'window-monitor-changed' at window tracker to signal that a window has moved from one monitor to another. This is done by tracking the last position and size to determine the old monitor where this window was placed. If the new position and size of this window is at another monitor this new signal is emitted.

This signal is used by the windows view to determine if the moved window has to stay in this view or if another window view instance has to create an actor for this window and to handle it.

This new signal is only useful in multi-monitor setups.

This commit addresses issue GH #88