Try not to eat up CPU ... again!
Do not emit 'window-geometry-changed' signals when application was suspended because there will be no visual changes. It is unneccessary to handle this signal.

Instead disconnect signal handler to 'geometry-changed' from windows at window tracker if the application is suspended and reconnect them if application resumes. Because we track the window's last position and size we emit a 'window-geometry-changed' for each window at application resume when it has been moved or resized in the meantime.

This should lower CPU usage while application is suspended and a window is moved or resized.

This commit addresses issue GH #44