Improve search responsiveness by delaying initial search
Instead of performing a search with the first keystroke we will now create a timeout source now. When the timeout is reached the initial search with the latest entered search term is performed. Further searches will not be delayed and performed immediately until the search term is empty or the search view is resetted.

The timeout interval to delay the initial is configurable by property "/components/search-view/delay-search-timeout" and expects an unsinged integer describe the timeout interval in milliseconds.

This commit addresses issue GH #86