Add experimental delayed window content creation
To make application "feel" faster on machines with slow graphics card (slow texture creation and/or uploading) a resume queue was implemented which will create the window texture in an idle source once at a time. This should help not to block the application at start-up or when switching to workspaces with many windows resulting in a "feeling" that the application is faster.

This implementation is experimental so a #define must be set and the debug message will be printed regardless if G_MESSAGES_DEBUG is set or not. Once the implementation is stable the message will printed with g_debug() instead of g_message().