Improve detection if tooltip should be shown
Previously the tooltip action only recognized to hide a tooltip if it was shown but it did not recognize not to show a tooltip if a interupting event (like button press) happened before the tooltip was shown at all. This commit fixes it and the tooltip should behave even more likely like a GTK+ tooltip now.