New experimental code to create and resume window content in an idle source is working well so make it configurable via a xfconf property called "/window-content-creation-priority".
The following values are valid: immediate, high, normal and low. The value 'immediate' will create the window content for each window at the time it was requested which is the default behaviour known so far. This value ('immediate') is also the default value if property was not set or set to an invalid value. The other values how fast and the number of window content textures should be created application is idle. While 'high' tries to create as many window content textures as possible, the values 'normal' and 'low' will create much less.