Implement function to allow a single-time switch to another view on resume as configured in settings or to switch to a view at initial start-up.
The view name set must be the internal view name of the view to switch to. This "switch" will only performed once, that means after the switch the name set will be resetted and the configured view in the settings will be considered again.

This commit is part of implementation needed to fix bug #12144