Rewrote view manager to use IDs.
Views do not specify their internal names by themselve but are now registered with an ID directly and the GType of this view is registered with this ID. Doing it this way makes it possible to define multi-purpose views of the same type. This is not interesting for views but for search providers.

Multi-purpose views (and later search providers) can reuse the same type but behave different depending on the ID. Thinking of one Gnome shell search provider object that will be registered several time but with different ID whereby the ID specifies with search provider to call via DBUS. This is just an example.