Implemented a basic and simple plugin system
This commit is huge as it implements a basic and simple plugin system as well a sample plugin adding a new view showing a clock when this plugin is enabled. By default the plugin manager does not load anything as it looks at the property "/enabled-plugin" in channel "xfdashboard" of xfconf which is empty by default. The plugin manager expects an array of string at this property whereby each string in this array contains the name of the plugin to load, e.g. "clock-view" for the sample clock view plugin.

The plugin are searched in the following search paths in this order: the path at the environment variable XFDASHBOARD_PLUGINS_PATH (if set), $HOME/.local/share/xfdashboard/plugins, (install prefix)/lib/xfdashboard/plugins.

The sample plugin can be tested by enabling it with the command (you should add parameter "-n" to xfconf-query at the beginning at the first time):
xfconf-query -c xfdashboard -p /enabled-plugins -a -s clock-view -t string

To disable all plugins use:
xfconf-query -c xfdashboard -r -p /enabled-plugins

Currently the plugin manager does not listen to changes at this property in xfconf so it does not load or unload any plugin automatically at changes. The application has to be restarted.