Re-add support for environment variable XFDASHBOARD_FORCE_NEW_INSTANCE
This environment variable can be set to any value as it needs only to exist. But it is only checked if xfdashboard is compiled with debug support.

If this enviroment variable is set to any value then it spawn a new instance of xfdashboard regardless if any other instance is already running. This is interesting to test a new compile of xfdashboard for debugging purposes while (a stable) xfdashboard instance is running in background in daemonized mode.