Allow mouse wheel to change workspace in windows view at any pointer position
The current behaviour of the setting "/components/windows-view/scroll-event-changes-workspace" is to change workspace if pointer is over the windows view (and only there) when the mouse wheel is used.

This behaviour was changed to allow the pointer to be anywhere on a monitor with an active windows view when the mouse wheel is used. Using the default theme that means that in a single monitor setup the windows view must be active to change workspace with mouse wheel (and enabled setting "/components/windows-view/scroll-event-changes-workspace" of course). The pointer can be anywhere: quicklaunch, search text box etc. In a multi-monitor setup the requirements of a single-monitor also affects the primary monitor in this setup but at all non-primary monitor is will always work if the pointer is at a non-primary monitor as it will only show a windows view ;)

This commit fixes GH #103 (and also some typos)