Implemented a new actor XfdashboardActionButton
The new actor XfdashboardActionButton is a button and derived from XfdashboardButton. It perform an action on a specified target actor class when clicked. It is configured similar to a (key) binding - see bindings.xml - and behaves similar but is reduced to a target and an action. E.g. if a XfdashboardActionButton is added to a layout (xfdashboard.xml) a target like "XfdashboardFocusManager" with action "focus-move-next" can be defined which will move the focus to next registered focusable actor like the default key "Tab" in bindings.xml

This implementation closes issues GH #102