Split function to determine desktop ID via environment variables from PID of newly opened windows in application tracker into operating system dependent functions.
The current implementation does not respect the operating system for which is was compiled. It is assumed that a proc filesystem does exist and provides a file at /proc/$PID/environ. But this is only true for Linux but not for *BSD systems. To support *BSD (if anyone using a BSD system would implement it) different operating system and their specific way to retrieve environment variables of a process the originate function was split into a independent and an OS-specified function. The OS-specific function for Linux was converted with this commit as the independent function was cleaned up. The OS-specific function will be compiled depending on which condition via #if defined(__XXX___) succeeds.

A fallback "OS-depending" function also exist just printed a warning once that this system is unsupported.