Check that fill canvas is set when allocating XfdashboardBackground actor.
At some circumstances that variable was not set (anymore?) and critical messages of clutter_canvas_set_size about CLUTTER_IS_CANVAS() failures were thrown. Maybe that XfdashboardBackground was already disposed which causes the fill canvas object to be destroyed but the allocation function was called anyway. Checking the pointer should silence this warnings ;)