Ensure that minimized flag is not set besides the maximized flags at window to consider it is maximized.
In my point of view it should be an error when the minimized flag is set along with the (both) flags for being maximized. But maybe it is just not a good name and it should be a 'hide' flag instead of 'minimized' flag. Then it makes sense that all these flags at set at once and we have to check additionally that the minimized (hide) flag is *not* set to consider that window is maximized.