Add hot-corner plugin
This is a plugin of the "other" category as it does not extend xfdashboard with any new views or search providers. It will check the pointer position and if it stays in an configured hot-corner (currently statically set to top-left corner) for a configured time interval (currently statically set to 350ms) it will switch the application's state. That means it will suspend the application if is active at that moment or activate it if was suspended. To do this xfdashboard must be running in daemon mode.