Ensure view is visible when allocation changes and scrollbar ranges might have changed.
Scrollbar ranges might get smaller, e.g. in search view when view is greater than viewpad and a new search is started. Then only the scrollbar ranges are adjusted to new size but view's clipping is not. This commit should ensure that the clipping is adjusted to position of scrollbars.

Hopefully fixes GH #112