Fix unpredictable reselections of result items in search view when the current selection was destroyed or hidden while updating results.
We check if the current selection is the first selectable item at the provider result container which contains the current selection _before_ updating the results. If it is the first selectable then remember the selection and the provider containing the selection. After updating the results we check if the current selection now is still the same as before updating result. If the selection has changed and we knew which provider contained the old selection then we select the first selectable item at provider's result container.

This commit should fix GH #119