Emit signal 'actor-created' at stage where a newly created actor was placed on.
This signal is emitted for the very first time a new actor gets its parent set and so it is the first time the actor belongs to UI and - from point of view of application - it exists now / was created. The main purpose of this signal is for plugins to react on newly created actor to filter and/or modify them and/or to connect signals. E.g. a plugin wants to connect a signal to each XfdashboardButton in XfdashboardQuicklaunch, so it has to listen for 'actor-created' signals at stage, then filter all actors for XfdashboardButtons whose parent is XfdashboardQuicklaunch. Then it can go on with its work.