Differ between "clicked" which is usually associated to a left-click and a "click" with any other button at XfdashboardClickAction.
Therefore the signals "clicked" and "long-press" at XfdashboardClickAction are only emitted if the click was done with the left button. To get clicks with any other the new signals "button-clicked" and "long-button-press" are emitted (also for left-clicks) which tranfer also the button which was clicked.

This changes the default behaviour of XfdashboardClickAction but should interfer with the current behaviour as left-clicks were assumed so far.