Next attempt to allow overriding draw function of background borders and corner, e.g. interesting for popup-menus to draw an arrow on any side of outline
This time the XfdashboardOutlineEffect provides a "draw" signal to paint the outline because the original code is the default signal handler.

Next XfdashboardBackground connects to this new signal of XfdashboardOutlineEffect to proxy this signal through itself. It also provides a virtual function (*draw_outline) which will be called if set. That function has to return TRUE if it handled the painting of outline and FALSE if not. If no virtual function was set, XfdashboardBackground will return FALSE to get the next signal handler to draw the outline getting called (usually the default signal handler of XfdashboardOutlineEffect).