Next attempt to allow overriding draw function of background borders and corner, e.g. interesting for popup-menus to draw an arrow on any side of background
This time XfdashboardBackground connects to "draw" signal of ClutterCanvas used to draw background (filled with a color and with rounded or not) and proxies this signal through itself. It also provides a virtual function (*draw_background) which will be called if set. That function has to return TRUE if it handled the painting of background canvas and FALSE if not. If no virtual function was set, XfdashboardBackground will return FALSE to get the next signal handler to draw the outline getting called (which is usually the default signal handler of XfdashboardBackground with the original code).

Also typos were fixed :)