Add fine-grained controllable debug system
If debug is enabled in configure script the user can use the environment variable XFDASHBOARD_DEBUG to enable categories to debug. But also classes like XfdashboardActor can be set in this environment variable to debug a specific class. Classes are case-sensitive and must be in camel-case. You can mix categories and classes in enviroment variable. You can set multiple categories and/or classes to debug by separating them by commas.