Revert "Revert the reversion "Add workarounds for known window managers having problems with default implementation""
This reverts commit 1b240ff21e31c788c9a74438e2e04621db0f26ce.

The workaround will not work regardless what I try. I do not know if I could blame the X11 backend of Clutter but it is strange that a second call to clutter_actor_show() will never show the stage in fullscreen although the size of the window seems to be correct for fullscreen.

I think other window managers have to wait for a GDK backend at xfdashboard. :(