Add compile-time checked support for xfconf version 4.12.x (called legacy here) and xfconf 4.13.x.
When using newer xfconf 4.13.x string arrays could not be retrieved anymore via a GPtrArray provided by dbus-glib but native GPtrArray gtype can be used. That means in addition that - when using xfconf 4.13.x - we can get rid of dbus-glib. So make dbus-glib optional and check for it only if xfconf 4.12.x was found and used in configure script

Thanks to Olivier Duchateau for the initial patch.

Fixes bug #13824