Fix function to lookup X11 window from GDK window for a ClutterStage in GDK backend
When xfdashboard is run in daemon mode the stage window will be created on first visible state. That means that looking up GDK window for a stage will return NULL at start-up until the stage is shown for the very first time. When iterating through the stages and determining GDK windows we need to respect this, otherwise xfdashboard crashes in daemon mode at start-up when resolving a NULL pointer of GDK window to a X11 window ID (XID).